+++ src/app/locales/en/components/file-browser/item-features-container.js
@@ -1,0 +1,3 @@
+export default {
+  expandTip: 'Show inherited features',
+};